+++ Moira/MoiraUtils.h
@@ -26,7 +26,7 @@ template<Size> u32 SUBS(u32 d1, u32 d2);
 template<Size> u64 MASK();
 template<Size> bool MSBIT(u32 data);
 template<Size> i32 SIGN(u32 data);
-template<Size> u8 BYTES();
+template<Size> u32 BYTES();
 template<Size S> bool ZERO(u32 data) { return CLIP<S>(data) == 0; }
 template<Size S> bool NEG(u32 data) { return SIGN<S>(data) < 0; }
 template<Size S> u32 SUBS(u32 d1, u32 d2) { return CLEAR<S>(d1) | CLIP<S>(d2); }
